requirements.txt
telethon==1.36.0
loguru==0.7.2
tqdm==4.66.4
python-dateutil==2.9.0